Move localstorage updates to action creators
